refactor: Place 내부 영업 시간 관련 객체 분리
