+++ backend/spring-routie/src/main/java/routie/place/controller/dto/response/PlaceReadResponse.java
@@ -23,10 +23,10 @@ public record PlaceReadResponse(
                 place.getName(),
                 place.getAddress(),
                 place.getStayDurationMinutes(),
-                place.getOpenAt(),
-                place.getCloseAt(),
-                place.getBreakStartAt(),
-                place.getBreakEndAt(),
+                place.getBusinessHour().getOpenAt(),
+                place.getBusinessHour().getCloseAt(),
+                place.getBusinessHour().getBreakStartAt(),
+                place.getBusinessHour().getBreakEndAt(),
                 place.getPlaceClosedDayOfWeeks().stream()
                         .map(PlaceClosedDayOfWeek::getClosedDayOfWeek)
                         .toList()
